Add files via upload
This code provides one with a bot that searches for the job roles. The job roles can be changed from within the code along with locations and login details through which the bot will automatically search through the roles. The bot search will be slow, which was intentional so that LinkedIn does not detect bot activity.

There is still an issue with the bot finding accurate results such as Latest job results and apply automatically to the most relevant ones.
However it is great with dealing with being undetected with bot activity as it works similar to a human and scrolls up and down and opens up description for various jobs.

These mentioned issues can be fixed with someone having higher knowledge as I used ChatGPT for the most part of this project. 

The program was on VS code and certain libraries has to be installed using CMD such as: 

>> pip install selenium webdriver-manager

>> pip install --user webdriver-manager

>> Installing chrome extension